soriting index cards by distance

+++ db/migrate/20220607125306_add_distance_to_toilets.rb
@@ -1,5 +1,5 @@
 class AddDistanceToToilets < ActiveRecord::Migration[7.0]
   def change
-    add_column :toilets, :toilet_distance, :integer
+    add_column :toilets, :toilet_distance, :float
   end
 end
